Move the module async to _async
