+++ gunicorn/workers/ggevent.py
@@ -27,7 +27,7 @@ from gevent import pywsgi
 
 import gunicorn
 from gunicorn.http.wsgi import base_environ
-from gunicorn.workers.async import AsyncWorker
+from gunicorn.workers._async import AsyncWorker
 from gunicorn.http.wsgi import sendfile as o_sendfile
 
 VERSION = "gevent/%s gunicorn/%s" % (gevent.__version__, gunicorn.__version__)
